requirements.txt
# Telegram Libraries
telethon>=1.34.0
aiogram>=3.4.0
pyrogram>=2.0.106

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
jinja2>=3.1.3
flask>=3.0.0
flask-cors>=4.0.0

# HTTP Client
httpx>=0.26.0

# Utilities
python-dotenv>=1.0.0
nest-asyncio>=1.5.0

# Performance Optimizations (Optional - requires C++ build tools on Windows)
# tgcrypto>=1.2.5  # Uncomment if you have Visual C++ Build Tools installed
uvloop>=0.19.0; sys_platform != 'win32'

# Note: 
# - uvloop is not available on Windows
# - tgcrypto requires Visual C++ Build Tools on Windows (optional, provides 2-4x speedup)
# - All other dependencies work on all platforms
# - The system works fine without tgcrypto, just slightly slower
